Add custom shared library location and change plugins' directory options
This first can be useful to alternate between AviSynth 2.5 and 2.6, or to
use AvsP without installing AviSynth by placing avisynth.dll/libavxsynth.so
on an AvsP's sudirectory.

Note that changing the autoload plugins directory writes to Registry and
changes the directory for all applications.  It's not an 'use an alternative
directory' like the other new option.  This is a limitation on AviSynth's
part (wish it used an environment variable as AvxSynth does).

Remove 'AvsP help directory' from Options.