Add reuse AviSynth environment option
This new option is intended for trying new filters/tweaking parameters.
Don't leave it on when not needed.  It's placed on the Video menu, off
by default.

Additionally, changing a slider now reuses the environment if the script
is unchanged, regardless of the preference state.

Advantages of using the same environment:
- Avoids scanning the autoload directory again after a change in the script.
  Noticeable speed-up for large directories.
- Variables are shared between all clips in the same environment. This may
  be used to avoid processing heavy filters again, e.g:

  LoadPlugin("heavy_filter.dll")
  source("video")
  clip = heavy_filter()
  clip.Tweak(cont=1.0)

With the new option checked, observing the effect of changing a Tweak
parameter can be speed-up by commenting the first lines or creating a
tagged section on them and turning it off:

  #LoadPlugin("heavy_filter.dll")
  #source("video")
  #clip = heavy_filter()
  clip.Tweak(cont=1.05)

or

  [heavy=0]
  LoadPlugin("heavy_filter.dll")
  source("video")
  clip = heavy_filter()
  [/heavy]
  clip.Tweak(cont=1.05)

The 'clip' variable is still in memory when the preview clip is recreated.
Note that the result would be incorrect if the same variable is assigned
again in the uncommented lines.  For that reason 'last' can't be used as
the saved variable.

Disadvantages:
- More memory usage
- Unexpected results if used without care