Change video slider behaviour when dragging the handle and the script is refreshed
Until now if the user changed with the mouse the position of the video
slider handle and the script needed to be refreshed the frame showed
afterwards was the one being originally shown, even if the frame count
is now different.

Change it so the position of the handle is kept, e.g. if a video originally
200 frames long is 500 after refresh and the position in which the handle
was released was #120, now it will be #300 (60% of the frame count for both
clips).