Add autocomplete filename menu option
This new feature is added to the existing 'autocomplete parameter' menu
entry.  If the characters are wrapped as a string literal (at least at the
beginning), this option tries to replace the whole string with a filename
using a shell-like auto-completion, using only the characters before the
cursor. Example:

'C:\work' directory:
    video_1.avi
    video_2.avi
    some_other_video.avi

Script (cursor at the end of 'vi'):
    AviSource("C:\work\vi
autocompletes to:
    AviSource("C:\work\video_
putting now the cursor at the start of the basename and typing 's':
    AviSource("C:\work\svideo_
autocompletes to:
    AviSource("C:\work\some_other_video.avi