Allow to save RGB48 data as 16-bit PNG
As AviSynth doesn't support 16-bit depth, each frame of the clip
returned by the script is assumed to be RGB48 data with a real
height half the reported by AviSynth. This can be achieved with
the following (Dither package):

Dither_convert_yuv_to_rgb (lsb_in=true, output="rgb48yv12")

or

Dither_convey_rgb48_on_yv12 (r, g, b)

As of wxPython 2.8 (stable) and 2.9.4 (development), additional
options can't be shown still in a 'save file' dialog. So, for now
instead of asking if the clip is 16-bit like for the JPEG quality
the script is parsed for one of the function calls above.

Added also a 'depth' parameter to the SaveImage macro function.