Add 'toggle preview placement' menu option
In addition to at the bottom, the video preview can be placed now to
the right (when not using a separate window).  Default is bottom, but
the current placement is also stored when saving a session and will
be set when loading the session file if it's not empty, including the
auto-saved startup session.

Closes #22