Don't import user function definitions whose function name starts with an underscore
This can be used to avoid importing auxiliar functions.  Note that
importing from avs is also used on start for populating the filter
database from the avsi files in the plugins directory if the option
is set in the program settings.

Of course no-one uses this convention because it doesn't mean anything
special to AviSynth.