Don't set the plugins autoload directory in HKCU when changing it
And to make sure the one in HKLM is used, delete the current value in HKCU.

While there's different 32 and 64 bit branches of HKLM\Software, that's not
the case for HKCU\Software. So until now, if the user had both AviSynth 32
and 64 bits installed and changed the plugins autoload directory for one of
them with AvsP, the directory for the other was changed accidentally to the
same value.